+++ app/domain/authentication/installed_authenticators.rb
@@ -4,6 +4,7 @@ module Authentication
   class InstalledAuthenticators
 
     AUTHN_RESOURCE_PREFIX = "conjur/authn-"
+    AUTHN_STATUS_FILTER = %r{conjur/(authn(?:-[^/]+)?(?:/[^/]+)?)$}
 
     class << self
       def authenticators(env, authentication_module: ::Authentication)
@@ -28,7 +29,7 @@ module Authentication
           .where(identifier.like("#{AUTHN_RESOURCE_PREFIX}%"))
           .where(kind => "webservice")
           .select_map(identifier)
-          .map { |id| id[%r{^conjur/(authn(?:-[^/]+)?(?:/[^/]+)?)$}, 1] } # filter out nested status webservice
+          .map { |id| id[AUTHN_STATUS_FILTER, 1] } # filter out nested status webservice
           .compact
           .push(::Authentication::Common.default_authenticator_name)
       end